requirements.txt
Flask
requests
playwright
python-dateutil
gunicorn 